+++ draftjs-compose/src/main/java/com/apoorvgupta/draftjscompose/utils/DraftJsConstants.kt
@@ -1,4 +1,4 @@
-package com.apoorvgupta.draftjs_compose.utils
+package com.apoorvgupta.draftjscompose.utils
 
 /**
  * @author Apoorv Gupta
