+++ draftjs-compose/src/main/java/com/apoorvgupta/draftjscompose/view/DraftJsView.kt
@@ -1,12 +1,12 @@
-package com.apoorvgupta.draftjs_compose.view
+package com.apoorvgupta.draftjscompose.view
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import com.apoorvgupta.draftjs_compose.data.DraftJS
-import com.apoorvgupta.draftjs_compose.utils.DraftJsUtils
+import com.apoorvgupta.draftjscompose.data.DraftJS
+import com.apoorvgupta.draftjscompose.utils.DraftJsUtils
 
 /**
  * @author Apoorv Gupta
